added handleAddSuccess
